Try innerblocks based details block
