--- packages/block-library/src/details-content/block.json
@@ -1,50 +1,0 @@
-{
-	"$schema": "https://schemas.wp.org/trunk/block.json",
-	"apiVersion": 2,
-	"__experimental": true,
-	"name": "core/details-content",
-	"title": "Details Content",
-	"category": "text",
-	"parent": [ "core/details" ],
-	"description": "Add content that may be shown or hidden via a Details block.",
-	"textdomain": "default",
-	"supports": {
-		"align": false,
-		"color": {
-			"gradients": true,
-			"link": true,
-			"__experimentalDefaultControls": {
-				"background": true,
-				"text": true,
-				"link": true
-			}
-		},
-		"__experimentalBorder": {
-			"radius": true,
-			"color": true,
-			"width": true,
-			"style": true
-		},
-		"html": false,
-		"lock": false,
-		"multiple": false,
-		"reusable": false,
-		"spacing": {
-			"margin": true,
-			"padding": true
-		},
-		"typography": {
-			"fontSize": true,
-			"lineHeight": true,
-			"__experimentalFontFamily": true,
-			"__experimentalFontWeight": true,
-			"__experimentalFontStyle": true,
-			"__experimentalTextTransform": true,
-			"__experimentalTextDecoration": true,
-			"__experimentalLetterSpacing": true,
-			"__experimentalDefaultControls": {
-				"fontSize": true
-			}
-		}
-	}
-}
