--- packages/block-library/src/details-content/save.js
@@ -1,12 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { useBlockProps, InnerBlocks } from '@wordpress/block-editor';
-export default function save() {
-	const blockProps = useBlockProps.save();
-	return (
-		<div { ...blockProps }>
-			<InnerBlocks.Content />
-		</div>
-	);
-}
