--- packages/block-library/src/details-summary/block.json
@@ -1,53 +1,0 @@
-{
-	"$schema": "https://schemas.wp.org/trunk/block.json",
-	"apiVersion": 2,
-	"__experimental": true,
-	"name": "core/details-summary",
-	"title": "Details Summary",
-	"category": "text",
-	"parent": [ "core/details" ],
-	"description": "Provide summary text used to toggle the display of content inside a Details block.",
-	"textdomain": "default",
-	"attributes": {
-		"summary": {
-			"type": "string"
-		}
-	},
-	"supports": {
-		"align": false,
-		"color": {
-			"gradients": true,
-			"__experimentalDefaultControls": {
-				"background": true,
-				"text": true
-			}
-		},
-		"__experimentalBorder": {
-			"radius": true,
-			"color": true,
-			"width": true,
-			"style": true
-		},
-		"html": false,
-		"lock": false,
-		"multiple": false,
-		"reusable": false,
-		"spacing": {
-			"margin": true,
-			"padding": true
-		},
-		"typography": {
-			"fontSize": true,
-			"lineHeight": true,
-			"__experimentalFontFamily": true,
-			"__experimentalFontWeight": true,
-			"__experimentalFontStyle": true,
-			"__experimentalTextTransform": true,
-			"__experimentalTextDecoration": true,
-			"__experimentalLetterSpacing": true,
-			"__experimentalDefaultControls": {
-				"fontSize": true
-			}
-		}
-	}
-}
