--- packages/block-library/src/details-summary/editor.scss
@@ -1,3 +1,0 @@
-.wp-block-details-summary div {
-	display: inline;
-}
