--- packages/block-library/src/details-summary/style.scss
@@ -1,3 +1,0 @@
-.wp-block-details-summary {
-	cursor: pointer;
-}
