+++ packages/block-library/src/details/editor.scss
@@ -1,0 +1,3 @@
+.wp-block-details summary div {
+	display: inline;
+}
