+++ packages/block-library/src/details/style.scss
@@ -1,3 +1,15 @@
 .wp-block-details {
+	box-sizing: border-box;
 	overflow: hidden;
+}
+
+// Use block gap for block; falls back to browser default if not supported.
+.wp-block-details > *:not(summary) {
+	margin-block-start: var(--wp--style--block-gap);
+	margin-block-end: 0;
+}
+
+// Remove excess margin from the last block.
+.wp-block-details > *:last-child {
+	margin-bottom: 0;
 }
