+++ packages/block-library/src/style.scss
@@ -12,7 +12,6 @@
 @import "./comment-template/style.scss";
 @import "./cover/style.scss";
 @import "./details/style.scss";
-@import "./details-summary/style.scss";
 @import "./embed/style.scss";
 @import "./file/style.scss";
 @import "./gallery/style.scss";
